+++ core/node/node.go
@@ -36,8 +36,8 @@ type Publisher interface {
 	Publish(topic string, data interface{})
 }
 
-// WebServer represents the web server for our web
-type WebServer interface {
+// UIServer represents the web server for our web
+type UIServer interface {
 	Serve() error
 	Stop()
 }
@@ -49,7 +49,7 @@ func NewNode(
 	publisher Publisher,
 	metricsSender *metrics.Sender,
 	natPinger NatPinger,
-	webServer WebServer,
+	uiServer UIServer,
 ) *Node {
 	return &Node{
 		connectionManager: connectionManager,
@@ -57,7 +57,7 @@ func NewNode(
 		publisher:         publisher,
 		metricsSender:     metricsSender,
 		natPinger:         natPinger,
-		webServer:         webServer,
+		uiServer:          uiServer,
 	}
 }
 
@@ -68,7 +68,7 @@ type Node struct {
 	publisher         Publisher
 	metricsSender     *metrics.Sender
 	natPinger         NatPinger
-	webServer         WebServer
+	uiServer          UIServer
 }
 
 // Start starts Mysterium node (Tequilapi service, fetches location)
@@ -87,9 +87,9 @@ func (node *Node) Start() error {
 	}
 
 	go func() {
-		err := node.webServer.Serve()
+		err := node.uiServer.Serve()
 		if err != nil {
-			log.Error("Web server error", err)
+			log.Error("UI server error", err)
 		}
 	}()
 
@@ -124,7 +124,7 @@ func (node *Node) Kill() error {
 	node.httpAPIServer.Stop()
 	log.Info("Api stopped")
 
-	node.webServer.Stop()
+	node.uiServer.Stop()
 	log.Info("Web server stopped")
 
 	node.natPinger.Stop()
