+++ ui/noop.go
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package web
+package ui
 
 import "sync"
